pyproject.toml
[project]
name = "stocks-mcp"
version = "0.1.0"
description = "Deploying an Stocks MCP server on Cloud Run"
requires-python = ">=3.13"
dependencies = [
    "fastmcp==2.12.4",
    "beautifulsoup4>=4.14.0",
    "pandas==2.3.3",
    "xmltodict==1.0.2",
    "sayou-stock>=0.1.0",
    "google-cloud-storage==3.5.0",
    "google-cloud-bigquery==3.38.0",
    "google-cloud-secret-manager==2.25.0",
    "lxml==6.0.2",
]
